- MFH: Suppress signedness warnings
git-svn-id: http://svn.php.net/repository/php/php-src/php/php-src/branches/PHP_5_3@263460 c90b9560-bf6c-de11-be94-00142212c4b1

+++ Zend/zend_builtin_functions.c
@@ -1808,7 +1808,7 @@ ZEND_FUNCTION(zend_test_func)
 #ifdef ZTS
 ZEND_FUNCTION(zend_thread_id)
 {
-	RETURN_LONG(tsrm_thread_id());
+	RETURN_LONG((long)tsrm_thread_id());
 }
 #endif
 #endif
